Extract SessionMaintainer interface to allow for different types of SessionMaintainers to run depending on charcteristics of the process that owns them.
